[IMP] website: review snippet tabs
Adding new options that are available through Bootstrap (Tabs or pills
class for nav-item. Alignment, menu horizontal/vertical, etc). Adding a
slide transition option. The snippet becomes also a drag&drop zone.

task-2152185
