+++ addons/website/static/src/scss/website.scss
@@ -626,7 +626,6 @@ table.table_desc tr td {
     }
 }
 
-
 //
 // Layout
 //
