+++ addons/website/views/assets.xml
@@ -114,6 +114,7 @@
     <script type="text/javascript" src="/website/static/src/snippets/s_popup/options.js"/>
     <script type="text/javascript" src="/website/static/src/snippets/s_product_catalog/options.js"/>
     <script type="text/javascript" src="/website/static/src/snippets/s_chart/options.js"/>
+    <script type="text/javascript" src="/website/static/src/snippets/s_tabs/options.js"/>
 
     <script type="text/javascript" src="/website/static/src/js/editor/wysiwyg_multizone.js"/>
     <script type="text/javascript" src="/website/static/src/js/editor/wysiwyg_multizone_translate.js"/>
